Imports correction, JWT project delete, add implementation

+++ app/src/api/projects.py
@@ -30,9 +30,13 @@ async def show(pagination:PaginationDep, request:Request):
         ProjectRelDTO.model_validate(row, from_attributes=True) for row in result])
 
 @router.post("/v1/projects") #Done
-async def add(project:ProjectCreateDTO):
+async def add(project:ProjectCreateDTO,request:Request):
     try:
-        user = await Users.find_one_or_none_by_id(project.owner_id)
+        if project.owner_id == 0:
+            user = await verify_token(request.cookies.get("auth_token"))
+            project.owner_id = user.id
+        else:
+            user = await Users.find_one_or_none_by_id(project.owner_id)
         if not user:
             raise HTTPException(status_code=404,
                                 detail="Пользователя для создания проекта не существует")
@@ -42,8 +46,12 @@ async def add(project:ProjectCreateDTO):
         raise HTTPException(status_code=500,detail=f"{_ex}")
     
 @router.delete("/v1/projects") #Done
-async def remove(project_id:int, owner_id:int):
+async def remove(project_id:int,request:Request):
     try:
+        user = await verify_token(request.cookies.get("auth_token"))
+        project = await Projects.find_one_or_none_by_id(project_id)
+        if project.owner_id != user.id:
+            raise HTTPException(status_code=401, detail="У вас нет прав на удаление проекта")
         res = await Projects.delete_by_id(project_id)
         return ResponseCreator.create_response(message="Проект успешно удален")
     except Exception as _ex:
